let tests work offline
Tests did not work when working offline.
This was because forestplot.dataframe_utils.load_data
and forestplot.tests.test_plot.py were trying to access csv's from
https://github.com/LSYS/forestplot/tree/main/examples/data.
To fix this, a option for setting an environment variable FORESTPLOT_OFFLINE=1
and an argument data_path to forestplot.dataframe_utils.load_data was added.
This makes it possible to provide a path to the directory containing csv's to load_data
and setting FORESTPLOT_OFFLINE=1 makes it fail immediately if unsuccesful.